Refactor: movement, and access changes
