+++ core/src/com/mrwalker/firstgame/Entity/Parts/Entity.java
@@ -3,10 +3,8 @@ package com.mrwalker.firstgame.Entity.Parts;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.math.Vector2;
 import com.mrwalker.firstgame.BodyID.BodyID;
-import com.mrwalker.firstgame.Entity.Parts.EntityBody.EntityBody;
 import com.mrwalker.firstgame.Entity.Models.Behaviour;
 import com.mrwalker.firstgame.Entity.Models.EntityState;
-import com.mrwalker.firstgame.Entity.Parts.Animations.EntityAnimation;
 import com.mrwalker.firstgame.auxiliary.Position2;
 
 import org.jetbrains.annotations.NotNull;
@@ -14,9 +12,9 @@ import org.jetbrains.annotations.NotNull;
 public class Entity {
     private static final String TAG = Entity.class.getSimpleName();
 
-    private final EntityBody body;
-    private final EntityState state;
-    private final EntityAnimation animation;
+    protected final EntityBody body;
+    protected final EntityState state;
+    protected final EntityAnimation animation;
 
     public Entity(@NotNull EntityState state,
                   @NotNull EntityBody body,
