+++ core/src/com/mrwalker/firstgame/Entity/Parts/EntityBody.java
@@ -1,4 +1,4 @@
-package com.mrwalker.firstgame.Entity.Parts.EntityBody;
+package com.mrwalker.firstgame.Entity.Parts;
 
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.physics.box2d.Body;
